+++ src/main/java/nextstep/subway/applicaion/line/LineModifyService.java
@@ -10,6 +10,9 @@ import nextstep.subway.exception.NotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.List;
+import java.util.Optional;
+
 @Service
 @Transactional
 public class LineModifyService {
@@ -34,6 +37,10 @@ public class LineModifyService {
         line.changeLineInformation(lineRequest.getName(), lineRequest.getColor());
     }
 
+    public void deleteLine(Long id) {
+        lineRepository.deleteById(id);
+    }
+
     public void addSection(Long id, Section section){
         Line line = lineRepository.findById(id).orElseThrow(NotFoundException::new);
         // TODO 이게 맞는건지 모르겠습니다. 차라리 생성자 주입으로 해서 넣는게 맞는건지...
@@ -41,7 +48,20 @@ public class LineModifyService {
         line.addSection(section);
     }
 
-    public void deleteLine(Long id) {
-        lineRepository.deleteById(id);
+    public void deleteSection(Long id, Long stationId){
+        Line line = lineRepository.findById(id).orElseThrow(NotFoundException::new);
+        List<Section> sections = line.getSections();
+
+        if(sections.size() <= 1){
+            throw new IllegalArgumentException();
+        }
+
+        Section lastSection = sections.get(sections.size() -1);
+        if(lastSection.getDownStation().getId() != stationId){
+            throw new IllegalArgumentException();
+        }
+
+        sections.remove(lastSection);
     }
+
 }
