+++ src/test/java/nextstep/subway/acceptance/LineAcceptanceTest.java
@@ -206,72 +206,4 @@ class LineAcceptanceTest extends AcceptanceTest {
                 .isEqualTo(HttpStatus.CONFLICT.value());
     }
 
-    @DisplayName("노선에 구간 추가")
-    @Test
-    void addNewStationSectionTest() {
-        // given
-        // 노선을 생성하고 해당 노선에 종점역을 추가한 후
-        ExtractableResponse<Response> createFirstStationResponse =
-          StationRequest.stationCreateRequest("강남역");
-        ExtractableResponse<Response> createSecondStationResponse =
-          StationRequest.stationCreateRequest("역삼역");
-        ExtractableResponse<Response> createThirdStationResponse =
-          StationRequest.stationCreateRequest("선릉역");
-
-        Long upStationId = createFirstStationResponse.jsonPath().getLong("id");
-        Long downStationId = createSecondStationResponse.jsonPath().getLong("id");
-        Long newDownStationId = createThirdStationResponse.jsonPath().getLong("id");
-
-        ExtractableResponse<Response> response =
-          lineCreateRequest(
-            LINE_NAME_A, LINE_COLOR_A, upStationId, downStationId, FIRST_DISTANCE);
-
-        String uri = response.header(LOCATION);
-
-        Map<String, Object> createRequest = new HashMap<>();
-        createRequest.put(UP_STATION_ID, downStationId);
-        createRequest.put(DOWN_STATION_ID, newDownStationId);
-        createRequest.put(DISTANCE, 3);
-
-        // when
-        ExtractableResponse<Response> newSectionResponse = RestAssured.given()
-          .log()
-          .all()
-          .body(createRequest)
-          .contentType(MediaType.APPLICATION_JSON_VALUE)
-          .when()
-          .post(uri + "/sections")
-          .then()
-          .log()
-          .all()
-          .extract();
-
-        // then
-        assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
-    }
-
-    @DisplayName("노선 구간 제거")
-    @Test
-    void deleteStationSectionTest() {
-        // given
-        // 노선을 생성하고 해당 노선에 구간을 추가한 후
-        ExtractableResponse<Response> createFirstStationResponse =
-          StationRequest.stationCreateRequest("강남역");
-        ExtractableResponse<Response> createSecondStationResponse =
-          StationRequest.stationCreateRequest("역삼역");
-
-        Long upStationId = createFirstStationResponse.jsonPath().getLong("id");
-        Long downStationId = createSecondStationResponse.jsonPath().getLong("id");
-
-        ExtractableResponse<Response> response =
-          lineCreateRequest(
-            LINE_NAME_A, LINE_COLOR_A, upStationId, downStationId, FIRST_DISTANCE);
-
-        // when
-        // 해당 구간을 삭제하면
-
-        // then
-        // 구간이 삭제된다.
-    }
-
 }
